♻️ refactor(quality-phase2): split 781-line SpecStepNav into focused modules
Pure extraction, no behavior change:
- SpecStepNav.tsx (135) — step navigation bar only; re-exports keep the
  existing import surface for SlideCarousel
- SpecMarkdownPreview.tsx (372) — spec content rendering + art-direction
  gallery/preview/result states + color swatches
- StyleCard.tsx (102) — style gallery card with iframe cover
- SpecWaiting.tsx (206) — Brief/Outline/ArtDirection waiting animations

Also removes a dangling 'export' fragment left mid-file before the
waiting-animations section.
